sudolver_api/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Andrin Meier", email = "andrin@pricemeier.ch"},
]
dependencies = [
    "numpy>=1.21.1",
    "opencv-python>=4.5.4.60",
    "ortools-puzzle-solvers>=0.0.6",
    "fastapi>=0.75.2",
    "uvicorn[standard]>=0.17.6",
    "python-dotenv>=0.20.0",
    "joblib>=1.1.0",
    "scikit-image>=0.19.2",
    "imgaug>=0.4.0",
    "scikit-learn>=1.0.2",
    "scipy>=1.6.1",
    "matplotlib>=3.5.2",
    "boto3>=1.22.10",
]
requires-python = ">=3.7"
license = {text = "MIT"}
[project.optional-dependencies]

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.3.0",
    "pytest>=7.1.2",
    "flake8>=4.0.1",
    "mypy>=0.950",
    "coverage>=6.3.3",
    "flake8-bandit>=3.0.0",
]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
lint = "flake8 ."
docker-lint = "hadolint Dockerfile"
format = "black ."
typecheck = "mypy app tests"
test = "pdm run -v pytest tests"
testci = "coverage run -m pytest -v tests"
testreport = "coverage xml"
dev = "uvicorn app.main:app --reload"